Move Bigtable CLI args out of shared KvArgs into sui-indexer-alt-jsonrpc
Bigtable-instance configuration only needs to live where it's still used.
Split it out of the shared KvArgs struct in sui-indexer-alt-reader into a
new BigtableKvArgs owned by sui-indexer-alt-jsonrpc, with no change to that
binary's CLI surface or behavior.
