+++ crates/sui-indexer-alt-e2e-tests/tests/jsonrpc_write_api_tests.rs
@@ -20,6 +20,7 @@ use sui_indexer_alt::setup_indexer;
 use sui_indexer_alt_framework::IndexerArgs;
 use sui_indexer_alt_framework::ingestion::ClientArgs;
 use sui_indexer_alt_framework::ingestion::ingestion_client::IngestionClientArgs;
+use sui_indexer_alt_jsonrpc::BigtableKvArgs;
 use sui_indexer_alt_jsonrpc::NodeArgs;
 use sui_indexer_alt_jsonrpc::RpcArgs;
 use sui_indexer_alt_jsonrpc::args::SystemPackageTaskArgs;
@@ -107,6 +108,7 @@ impl WriteTestCluster {
             Some(database_url),
             DbArgs::default(),
             KvArgs::default(),
+            BigtableKvArgs::default(),
             ConsistentReaderArgs::default(),
             RpcArgs {
                 rpc_listen_address,
