+++ crates/sui-indexer-alt-jsonrpc/Cargo.toml
@@ -62,6 +62,7 @@ sui-indexer-alt-reader.workspace = true
 sui-indexer-alt-schema.workspace = true
 sui-json.workspace = true
 sui-json-rpc-types.workspace = true
+sui-kvstore.workspace = true
 sui-name-service.workspace = true
 sui-open-rpc.workspace = true
 sui-open-rpc-macros.workspace = true
