+++ crates/sui-indexer-alt-jsonrpc/src/main.rs
@@ -44,6 +44,7 @@ async fn main() -> anyhow::Result<()> {
             database_url,
             db_args,
             kv_args,
+            bigtable_kv_args,
             consistent_reader_args,
             rpc_args,
             system_package_task_args,
@@ -76,6 +77,7 @@ async fn main() -> anyhow::Result<()> {
                 Some(database_url),
                 db_args,
                 kv_args,
+                bigtable_kv_args,
                 consistent_reader_args,
                 rpc_args,
                 node_args,
